Refactor stress bot for improved command handling and tests
